==> code/02_scraping_reports.ipynb <==
# Scraping reports

# import libs 

import os 
import fnmatch # file name matching
import tabula # extract table objects from PDF files
import camelot # tabula alternative

# os.getcwd()

## Setup

# check the current working directory 

# cp *NSLVE* ./nslve

os.chdir("/home/jae/scraping_slsv/outputs/nslve")

cwd = os.getcwd()

# print the cwd 
print("Current working directory: {0}". format(cwd))

# list files 

file_names = os.listdir(cwd)

# placeholder 

nslve_files = []

# for loop

for file in file_names:
    if fnmatch.fnmatch(file.lower(), "*nslve*"):
        nslve_files.append(file)

# check 

nslve_files[0:4]

# load template 

temp = []

for file in file_names:
    if fnmatch.fnmatch(file, "*json"):
        temp.append(file)

# check
os.getcwd() + "/" + ''.join(temp)

## Extract Table obejcts from PDF files

df = camelot.read_pdf(filepath = os.getcwd() + "/" + nslve_files[0],
                      pages = 'all')

for i in range(df.n):
    print(df[i])

df = tabula.read_pdf_with_template(input_path = os.getcwd() + "/" + nslve_files[0],
                              #batch = os.getcwd(),
                              template_path = os.getcwd() + "/" + ''.join(temp),
                              stream = True,
                              guess = False,
                              pages = "all")

tabula.convert_into_by_batch(os.getcwd(), 
                             output_format = 'csv',
                             pages = 'all',
                             stream = True)
