==> code/03_cleaning_merging.Rmd <==
---
title: "Cleaning and Merging"
author: "Jae Yeon Kim"
date: "`r Sys.Date()`"
output: html_document
---

# Install pkgs 

```{r}
if(!require(pacman)) install.packages("pacman")
pacman::p_load(tidyverse, here, glue, purrr, stringr, 
               xlsx, # import xlsx file  
               tm, # text processing
               stringdist, # fuzzy matching
               fuzzyjoin,
               fedmatch,
               corrr, 
               scales,
               patchwork,
               boot, 
               purrr
               )

source(here("functions", "utils.R"))
```

# Load data 

```{r}
# ap = action plans
ap <- read_csv(here("outputs", "filtered_df.csv"))

paste("The number of action plans:", nrow(ap))

meta <- read.xlsx(here("raw_data", list.files(here("raw_data"))), sheetIndex = 1)
```

# Wrangle data 

## Score cards 

```{r}
sc <- data.frame(name = tolower(meta[2:nrow(meta),1]))

sc <- sc %>%
         mutate(name = str_replace_all(name, "[:punct:]", ""),
                name = trimws(name)) 
```

```{r}
test_matching <- function(method_name) {

  matched <- stringdist_join(
    sc, 
    ap,
    by = c("name" = "name"),
    method = method_name,
    max_dist = 2,
    distance_col = "distance"
  ) 

  dist <- nrow(sc) -
  matched$name.x %>% unique() %>% length()
  
  out <- data.frame(method = method_name, dist = dist)
  
  return(out)

}

test_out <- map_dfr(c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"), test_matching) %>%
  arrange(dist)

test_out
```

```{R}
ac_high <- meta[2:nrow(meta),c(3:5)] %>%
  mutate(accountability = "high") 
ac_low <- meta[2:nrow(meta),c(6:8)] %>%
  mutate(accountability = "low")

l_high <- meta[2:nrow(meta),c(9:11)] %>%
  mutate(learning = "high")
l_low <- meta[2:nrow(meta),c(12:14)] %>%
  mutate(learning = "low")

sc_high <- meta[2:nrow(meta),c(15:17)] %>%
  mutate(strategic_capacity = "high")
sc_low <- meta[2:nrow(meta),c(17:19)] %>%
  mutate(strategic_capacity = "low")
```

```{r}
bind_2df <- function(df1, df2) {

  names(df1)[1:3] <- c("cvp", "slsv", "all_in")
  names(df2)[1:3] <- c("cvp", "slsv", "all_in")
  
  df1 <- df1 %>%
    rownames_to_column("id")
  df2 <- df2 %>%
    rownames_to_column("id")
  
  binded <- bind_rows(df1, df2)

  binded[is.na(binded)] <- 0
  
  out <- binded %>%
    pivot_longer(last_col(),
               names_to = "type",
               values_to = "score") %>%
    mutate(name = rep(sc$name, 2)) %>%
    select(name, everything()) 
  
  return(out)
}
```

```{r}
sc_binded <- bind_rows(bind_2df(ac_high, ac_low),
  bind_2df(l_high, l_low)) %>%
  bind_rows(bind_2df(sc_high, sc_low))

write_csv(sc_binded, here("processed_data", "sc_binded.csv"))
```

```{r}
sc_binded <- sc_binded %>%
    mutate(
      cvp = as.numeric(cvp),
      slsv = as.numeric(slsv),
      all_in = as.numeric(all_in)
      )

sc_binded <- sc_binded %>% 
  mutate(slsv = replace_na(slsv, 0))
```

```{r}
sc_binded %>%
  mutate(type = recode(type,
      accountability = "Accountability",
      learning = "Learning",
      strategic_capacity = "Strategic capacity"
                       )) %>%
  pivot_longer(cols = c("cvp","slsv","all_in"),
               names_to = "coder",
               values_to = "label") %>%
  mutate(coder = recode(coder, 
      cvp = "Campus Vote Project",
      slsv = "SLSV Coalition",
      all_in = "ALL IN Campus Democracy Challenge")) %>%
  mutate(score = recode(score,
      high = "High",
      low = "Low")) %>%
  group_by(type, score, coder) %>%
  summarize(avg_labeled = mean(label)) %>%
  ggplot(aes(x = coder, y = avg_labeled, fill = score)) +
  geom_col(position = "dodge") +
  geom_text(aes(label = paste0(round(avg_labeled, 2)*100,"%")), position = position_dodge(width = .9)) +
  facet_wrap(~type) +
  coord_flip() +
  labs(x = "",
       y = "% labeled", 
       fill = "Score") +
  scale_y_continuous(label = scales::percent) +
  ggthemes::theme_economist()

ggsave(here("figures", "coder_label_dist.png"),
       height = 8,
       width = 9)
```

```{r}
sc_binded[,c(3:5)][is.na(sc_binded[,c(3:5)])] <- 0

plot_cor <- function(score_label, type_label) {
  
  out <- sc_binded %>%
    filter(score == score_label) %>%
    filter(type == type_label) %>%
    select(cvp, slsv, all_in) %>%
    correlate() %>% 
    rearrange() %>%
    autoplot() +
    geom_text(aes(label = round(r, digits = 2)), size = 4) +
    theme_bw(16) +
    labs(title = glue("{score_label} {type_label}"))
  
  return(out)
  
}

vars <- expand.grid(c("high", "low"), c("accountability", "learning", "strategic_capacity"))

var1 <- vars[,1]
var2 <- vars[,2]

glue("plot_cor('{var1}', '{var2}')")

#plot_cor('low', 'accountability')
#plot_cor('low', 'learning')

plot_cor('high', 'accountability') +
plot_cor('high', 'learning') +
plot_cor('high', 'strategic_capacity') +
plot_cor('low', 'strategic_capacity')

ggsave(here("figures", "corr_type.png"),
       height = 10,
       width = 10)
```

## Action plans 

```{r}
# clean text
stopwords_removal <- stopwords("en")  # Specify the language for stopwords (e.g., "en" for English)

ap$text <- tm::removeWords(ap$text, words = stopwords_removal)
```

```{r}
ap <- ap %>%
  mutate(name = tolower(name)) %>%
  mutate(name = str_replace_all(name, "-|action|plan", " "),
         name = str_replace_all(name, ".pdf", ""),
         name = str_replace_all(name, "\\d+", ""), 
         name = str_replace_all(name, "final", ""),
         name = str_replace_all(name, "may|june|docx|updated|dec|april|vfc|draft|update|revised|revise|jan|july|aug|mar", ""),
         name = str_replace_all(name, "[:punct:]", ""),
         name = str_replace_all(name, "all in challenge|spring", ""),
         name = trimws(name)) %>%
  select(name, text) 
```

# Join the action plans and the score cards

```{r}
ap <- ap %>%
  rownames_to_column("id")
```

```{r}
sc_high_df <- sc_binded %>%
  filter(score == "high" & type == "strategic_capacity")

sc_high_df <- sc_high_df %>%
  mutate(high_sc = ifelse(cvp == 1 | slsv == 1 | all_in == 1, 1, 0))

sc_high_df <- sc_high_df %>%
  select(name, id, high_sc)
```

```{r}
matched <- stringdist_left_join(
    sc_high_df %>% select(-id), 
    ap %>% select(-id),
    by = c("name" = "name"),
    method = "osa",
    max_dist = 1,
    distance_col = "distance"
  )

matched %>%
  filter(distance != 0) %>%
  select(name.x, name.y, distance, high_sc)
```

# Summarize the meta data 

```{r}
matched <- matched %>%
  distinct() %>%
  select(name.x, name.y, everything())
```

```{r}
write_csv(matched, here("processed_data", "matched.csv"))
```

```{r}
set.seed(1234)

word_cnt_dist <- matched %>%
  group_by(high_sc) %>%
  nest() %>% 
  mutate(boot_res = map(data,
                        ~ boot(data = stringr::str_count(.$text, '\\w+'),
                               statistic = function(x, i) mean(x[i], na.rm = T),
                               R = 1000)),
         boot_res_ci = map(boot_res, boot.ci, type = "perc"),
         mean = map(boot_res_ci, ~ .$t0),
         lower_ci = map(boot_res_ci, ~ .$percent[[4]]),
         upper_ci = map(boot_res_ci, ~ .$percent[[5]]),
         n =  map(data, nrow)) %>%
  select(-data, -boot_res, -boot_res_ci) %>% 
  unnest(cols = c(n, mean, lower_ci, upper_ci)) %>% 
  ungroup()

word_cnt_dist$mean[1] - word_cnt_dist$mean[2]
```

```{r}
word_cnt_dist %>%
  mutate(high_sc_fac = ifelse(high_sc == 1, "High strategic capacity", "Not")) %>%
  ggplot(aes(x = high_sc_fac, y = mean, 
             ymax = upper_ci, 
             ymin = lower_ci)) +
  geom_col() +
  geom_text(aes(label = round(mean, 2)), col = "red") +
  geom_errorbar() +
  ggthemes::theme_economist() +
  labs(x = "", y = "Word count")

ggsave(here("figures", "word_cnt.png"))
```
